# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movie Api Documentation",
        "description": "From L5 Swagger",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0."
    },
    "servers": [
        {
            "url": "http://movieapi.local",
            "description": "Develop API Server"
        }
    ],
    "paths": {
...
        "/api/movie/detail/{serverId}": {
            "get": {
                "tags": [
                    "movie"
                ],
                "summary": "Get movie detail by serverId",
                "operationId": "893bfb5604ab9ca10da60a217117cb44",
                "parameters": [
                    {
                        "name": "serverId",
                        "in": "path",
                        "description": "Server id of movie",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    }
}